[최건] commit message: refactor: Schedule 도메인 패키지 이동 및 기능 개선 - Schedule 도메인 클래스들을 project에서 team 패키지로 이동     - Schedule 엔티티 및 관련 DTO 이동     - Repository 및 Service 클래스 패키지 경로 수정     - import 문 업데이트
- Schedule 기능 개선
    - ScheduleRequestDto에 status 필드 추가
    - update 메서드에 status 업데이트 로직 추가
    - 알림 발송 후 isAlertSent 상태 업데이트

- 코드 정리
    - 불필요한 주석 제거
    - 코드 포맷팅 개선
    - 미사용 import 제
