+++ src/main/java/com/beyond/backend/domain/team/dto/AlertResponseDto.java
@@ -1,4 +1,4 @@
-package com.beyond.backend.domain.project.dto;
+package com.beyond.backend.domain.team.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
