+++ src/main/java/com/beyond/backend/domain/team/repository/ScheduleRepository.java
@@ -1,6 +1,6 @@
 package com.beyond.backend.domain.team.repository;
 
-import com.beyond.backend.domain.project.entity.Schedule;
+import com.beyond.backend.domain.team.entity.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long>, ScheduleRepositoryCustom{
